restructure the Steps.js file with the new way of doing things
