+++ client/elife-theme/src/elements/ui/index.js
@@ -3,3 +3,4 @@ export { default as Checkbox } from './Checkbox'
 export { default as Heading } from './Heading'
 export { default as TextField } from './TextField'
 export * from './Heading'
+export * from './Steps'
